iot_module/requirements.txt
Flask-Login
Flask==3.0.0
requests==2.31.0
pymongo==3.12.0
paho-mqtt


